Add music copyright info and update CV
The copyright is for Kevin Macleod, I used "March of the Spoons" in
Scaly Sewers Adventures. I need to give credit as specified by the CC
license here, since I am no longer linking to a YT video (with credits
in the description), and since I do not provide the game, so there is
nowhere else to see the copyright notice.