Link to email, put email in full
I will take the risk of bots scraping this e-mail for now, and put a
one-click mailto: link in the homepage. Also replaced 'AT' and 'DOT'
with '@' and '.' so the entire email appears consecutively, but with
HTML comments sprinkled throughout to deter bots. Although, the comments
probably do not do much since the mailto: link is right there.