Fix outerHTML on no-parent elem. for non-firefox
Chrome and Edge was erroring due to a weird behaviour with setting the
outerHTML. The code worked before adding viewport html tags to other
html files, but now could not reproduce the successful state from
earlier? Decided to set inner HTML of empty tags instead. This is
probably a better deterrent, anyways.