Try buttons navbar, change background and internal link colour
The old internal link colour was hard to see (esp. for protanopia
colour blindness). This new colour will hopefully be visible in
constrast with the darker background, too.

Changed the navbar to a more roundish, button-y look. May change it to
be more or less round, or do something else altogether in the future.

Most of these changes are courtesy of tips from a friend of a friend
(who is staying anonymous).