Update timeline with dog care
I am no longer actively looking for a job, while I am caring for an old
dog. I expect this period of time to last for a few more months from
this commit... but that estimation may be wrong.

In the mean time, I will look for opportunities later in the year.